move out create config stub
