+++ src/test/configuration.test.ts
@@ -2,15 +2,9 @@ import * as vscode from "vscode";
 import * as assert from "assert";
 import * as sinon from "sinon";
 
-import {
-  defaultCategories,
-  defaultSortOrder,
-  defaultPseudoSortOrder,
-  defaultCustomPrefixes,
-  defaultSortOnSave,
-} from "../lib/defaultConfig";
-import getClassesMap from "../getClassesMap";
 import { sortOnSave } from "../extension";
+import getClassesMap from "../getClassesMap";
+import createConfigStub from "./_createConfigStub";
 
 suite("VS Code Configuration", () => {
   teardown(() => {
@@ -153,41 +147,3 @@ suite("VS Code Configuration", () => {
     assert.strictEqual(sortedTailwind.includes("grid hover:grid"), true);
   });
 });
-
-function createConfigStub(options = {}) {
-  const defaults = {
-    categories: defaultCategories,
-    categoryOrder: { sortOrder: defaultSortOrder },
-    pseudoClassesOrder: { sortOrder: defaultPseudoSortOrder },
-    customPrefixes: defaultCustomPrefixes,
-    sortOnSave: defaultSortOnSave,
-  };
-
-  const config = { ...defaults, ...options };
-
-  return sinon
-    .stub(vscode.workspace, "getConfiguration")
-    .callsFake((section) => {
-      if (section === "tailwindSorter") {
-        return {
-          get: (key: string) => {
-            switch (key) {
-              case "categories":
-                return config.categories;
-              case "categoryOrder":
-                return config.categoryOrder;
-              case "pseudoClassesOrder":
-                return config.pseudoClassesOrder;
-              case "customPrefixes":
-                return config.customPrefixes;
-              case "sortOnSave":
-                return config.sortOnSave;
-              default:
-                return undefined;
-            }
-          },
-        } as vscode.WorkspaceConfiguration;
-      }
-      return {} as vscode.WorkspaceConfiguration;
-    });
-}
